UI left

+++ app/src/main/java/ba/etf/rma23/projekat/UserRating.kt
@@ -3,5 +3,5 @@ package ba.etf.rma23.projekat
 data class UserRating(
     override val userName: String,
     override val timestamp: Long,
-    val rating: Double
+    val rating: Int
 ): UserImpression()
